Move completeness phase to folder
