+++ client/src/components/application/phases/completeness/CompletenessPhase.test.tsx
@@ -43,7 +43,7 @@ vi.mock("components/application/date/dateQueries", () => ({
 
 const mockCompletePhase = vi.fn();
 const mockDeclareCompletenessPhaseIncomplete = vi.fn();
-vi.mock("../phase-status/phaseCompletionQueries", () => ({
+vi.mock("../../phase-status/phaseCompletionQueries", () => ({
   useCompletePhase: vi.fn(() => ({
     completePhase: mockCompletePhase,
   })),
