+++ client/src/components/application/phases/completeness/index.ts
@@ -1,0 +1,1 @@
+export * from "./CompletenessPhase";
